refactor: split test files to one test per file per repo convention
