+++ tests/pcb/pcb-silkscreen-text-knockout-anchors1.test.ts
@@ -137,99 +137,3 @@ test("silkscreen knockout text - all 9-point anchor alignments", () => {
     "silkscreen-knockout-all-anchors",
   )
 })
-
-test("silkscreen knockout text - center anchor (default)", () => {
-  const centerAnchor = convertCircuitJsonToPcbSvg([
-    {
-      type: "pcb_board",
-      pcb_board_id: "pcb_board_0",
-      width: 15,
-      height: 10,
-      center: { x: 0, y: 0 },
-      num_layers: 2,
-      material: "fr4",
-      thickness: 1.2,
-    },
-    {
-      type: "pcb_silkscreen_text",
-      layer: "top",
-      pcb_silkscreen_text_id: "text_center_default",
-      font: "tscircuit2024",
-      font_size: 1,
-      pcb_component_id: "comp_0",
-      anchor_position: { x: 0, y: 0 },
-      anchor_alignment: "center",
-      text: "CENTER",
-      is_knockout: true,
-    },
-  ])
-
-  expect(centerAnchor).toMatchSvgSnapshot(
-    import.meta.path,
-    "silkscreen-knockout-center-anchor",
-  )
-})
-
-test("silkscreen knockout text - top_left anchor", () => {
-  const topLeftAnchor = convertCircuitJsonToPcbSvg([
-    {
-      type: "pcb_board",
-      pcb_board_id: "pcb_board_0",
-      width: 20,
-      height: 10,
-      center: { x: 0, y: 0 },
-      num_layers: 2,
-      material: "fr4",
-      thickness: 1.2,
-    },
-    {
-      type: "pcb_silkscreen_text",
-      layer: "top",
-      pcb_silkscreen_text_id: "text_top_left",
-      font: "tscircuit2024",
-      font_size: 1,
-      pcb_component_id: "comp_0",
-      anchor_position: { x: -5, y: -2 },
-      anchor_alignment: "top_left",
-      text: "TOPLEFT",
-      is_knockout: true,
-    },
-  ])
-
-  expect(topLeftAnchor).toMatchSvgSnapshot(
-    import.meta.path,
-    "silkscreen-knockout-top-left-anchor",
-  )
-})
-
-test("silkscreen knockout text - bottom_right anchor", () => {
-  const bottomRightAnchor = convertCircuitJsonToPcbSvg([
-    {
-      type: "pcb_board",
-      pcb_board_id: "pcb_board_0",
-      width: 20,
-      height: 10,
-      center: { x: 0, y: 0 },
-      num_layers: 2,
-      material: "fr4",
-      thickness: 1.2,
-    },
-    {
-      type: "pcb_silkscreen_text",
-      layer: "top",
-      pcb_silkscreen_text_id: "text_bottom_right",
-      font: "tscircuit2024",
-      font_size: 1,
-      pcb_component_id: "comp_0",
-      anchor_position: { x: 5, y: 2 },
-      anchor_alignment: "bottom_right",
-      text: "BOTRIGHT",
-      is_knockout: true,
-    },
-  ])
-
-  expect(bottomRightAnchor).toMatchSvgSnapshot(
-    import.meta.path,
-    "silkscreen-knockout-bottom-right-anchor",
-  )
-})
