+++ tests/pcb/pcb-silkscreen-text-knockout-combined2.test.ts
@@ -1,0 +1,35 @@
+import { test, expect } from "bun:test"
+import { convertCircuitJsonToPcbSvg } from "lib"
+
+test("silkscreen knockout text - 180° rotation with various anchors", () => {
+  const rotated180 = convertCircuitJsonToPcbSvg([
+    {
+      type: "pcb_board",
+      pcb_board_id: "pcb_board_0",
+      width: 25,
+      height: 15,
+      center: { x: 0, y: 0 },
+      num_layers: 2,
+      material: "fr4",
+      thickness: 1.2,
+    },
+    {
+      type: "pcb_silkscreen_text",
+      layer: "top",
+      pcb_silkscreen_text_id: "text_180",
+      font: "tscircuit2024",
+      font_size: 1,
+      pcb_component_id: "comp_0",
+      anchor_position: { x: 0, y: 0 },
+      anchor_alignment: "center",
+      text: "FLIPPED",
+      is_knockout: true,
+      ccw_rotation: 180,
+    },
+  ])
+
+  expect(rotated180).toMatchSvgSnapshot(
+    import.meta.path,
+    "silkscreen-knockout-180-rotation",
+  )
+})
