+++ tests/pcb/pcb-silkscreen-text-knockout-multiline1.test.ts
@@ -1,0 +1,34 @@
+import { test, expect } from "bun:test"
+import { convertCircuitJsonToPcbSvg } from "lib"
+
+test("silkscreen knockout text - multi-line text", () => {
+  const multiline = convertCircuitJsonToPcbSvg([
+    {
+      type: "pcb_board",
+      pcb_board_id: "pcb_board_0",
+      width: 25,
+      height: 15,
+      center: { x: 0, y: 0 },
+      num_layers: 2,
+      material: "fr4",
+      thickness: 1.2,
+    },
+    {
+      type: "pcb_silkscreen_text",
+      layer: "top",
+      pcb_silkscreen_text_id: "text_multiline",
+      font: "tscircuit2024",
+      font_size: 0.8,
+      pcb_component_id: "comp_0",
+      anchor_position: { x: 0, y: 0 },
+      anchor_alignment: "center",
+      text: "LINE1\nLINE2\nLINE3",
+      is_knockout: true,
+    },
+  ])
+
+  expect(multiline).toMatchSvgSnapshot(
+    import.meta.path,
+    "silkscreen-knockout-multiline",
+  )
+})
